+++ CHANGELOG.md
@@ -10,6 +10,7 @@ CHANGELOG
 - fix saving invalid data if Pendant of Reflection or Golden Goose equipped (issue #17)
 - fix showing hero as changed upon opening for some game versions
 - fix file extension filter in files list getting narrowed after deleting file
+- retain button focus on skills page properly over removing or reordering
 
 
 3.0, 2025-04-09